PlayerFacade tests

+++ GamingManagement-service/src/main/java/cz/fi/muni/pa165/yellow_yak/facade/PlayerFacadeImpl.java
@@ -30,30 +30,45 @@ public class PlayerFacadeImpl implements PlayerFacade {
 
     @Override
     public PlayerDTO create(String name, String email) {
+        if (name == null || email == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("creating player, name = {}, email = {}", name, email);
         return beanMappingService.mapTo(playerService.create(name, email), PlayerDTO.class);
     }
 
     @Override
     public void remove(Long id) {
+        if (id == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("removing player, id = {}", id);
         playerService.remove(id);
     }
 
     @Override
     public PlayerDTO findById(Long id) {
+        if (id == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("finding player by ID, id = {}", id);
         return beanMappingService.mapTo(playerService.findById(id), PlayerDTO.class);
     }
 
     @Override
     public List<PlayerDTO> findByUsername(String username) {
+        if (username == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("listing player by username, username = {}", username);
         return beanMappingService.mapTo(playerService.findByUsername(username), PlayerDTO.class);
     }
 
     @Override
     public List<PlayerDTO> findByTeam(Long teamId) {
+        if (teamId == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("listing player by team, teamId = {}", teamId);
         return beanMappingService.mapTo(playerService.findByTeam(teamId), PlayerDTO.class);
     }
